Create shared utility method for project questions
This currently only contains validation - as the form data for appointment
outcome and course completion form is slightly different.
